ADF_Profiler/requirements.txt
requests==2.32.3
azure-identity==1.24.0